+++ nflow-engine/src/main/java/io/nflow/engine/service/WorkflowInstanceService.java
@@ -56,8 +56,21 @@ public class WorkflowInstanceService {
    * @throws NflowNotFoundException If workflow instance is not found.
    */
   public WorkflowInstance getWorkflowInstance(long id, Set<WorkflowInstanceInclude> includes, Long maxActions) {
+    return getWorkflowInstance(id, includes, maxActions, false);
+  }
+
+  /**
+   * Return the workflow instance matching the given id.
+   * @param id Workflow instance id.
+   * @param includes Set of properties to be loaded.
+   * @param maxActions Maximum number of actions to be loaded.
+   * @param queryArchive Query archive if not found from main tables.
+   * @return The workflow instance
+   * @throws EmptyResultDataAccessException if not found
+   */
+  public WorkflowInstance getWorkflowInstance(long id, Set<WorkflowInstanceInclude> includes, Long maxActions, boolean queryArchive) {
     try {
-      return workflowInstanceDao.getWorkflowInstance(id, includes, maxActions);
+      return workflowInstanceDao.getWorkflowInstance(id, includes, maxActions, queryArchive);
     } catch (EmptyResultDataAccessException e) {
       throw new NflowNotFoundException("Workflow instance", id, e);
     }